Break NLP tutorial into parts
