+++ .gitignore
@@ -6,7 +6,12 @@ advanced
 #data things
 beginner_source/hymenoptera_data
 intermediate_source/data/
+advanced_source/images/
 *data.zip
+
+#builds
+_build/
+_static/thumbs/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
